tasks/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.1
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.22
pytz==2023.3
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
boto3==1.34.0
requests==2.31.0
tabulate==0.9.0
